added testcontainers tests
