+++ pom.xml
@@ -18,6 +18,12 @@
     <dependencies>
         <dependency>
             <groupId>org.testcontainers</groupId>
+            <artifactId>mongodb</artifactId>
+            <version>1.15.1</version>
+            <scope>test</scope>
+        </dependency>
+        <dependency>
+            <groupId>org.testcontainers</groupId>
             <artifactId>testcontainers</artifactId>
             <version>1.15.1</version>
             <scope>test</scope>
