+++ src/main/java/ru/eatthefrog/hatterBot/LoginManager/LoginInstanceImpl.java
@@ -6,19 +6,22 @@ import org.springframework.stereotype.Component;
 import ru.eatthefrog.hatterBot.MD5StringHasher.MD5StringHasher;
 
 public class LoginInstanceImpl implements LoginInstance {
+
     @Autowired
     MD5StringHasher md5StringHasher;
 
-    public String login;
-    public String password;
+    private String login;
+    private String password;
     public long lastTimeVerified;
     public Boolean isValid = false;
     public int verificationFrequency = 3600000;
     public LoginInstanceImpl(){};
+
     public LoginInstanceImpl(String loginTemp, String passwordTemp) {
-        login = loginTemp;
-        password = passwordTemp;
+        setLogin(loginTemp);
+        setPassword(passwordTemp);
     }
+
     public void setValid() {
         isValid = true;
         lastTimeVerified = System.currentTimeMillis();
@@ -32,20 +35,19 @@ public class LoginInstanceImpl implements LoginInstance {
     public void makeUltraOld() {
         lastTimeVerified = 0;
     }
+
     public void setNotValid() {
         isValid = false;
         lastTimeVerified = 0;
     }
+
     public Boolean isItTimeToVerify() {
         return System.currentTimeMillis() - lastTimeVerified > verificationFrequency;
     }
 
     @Override
     public String getPasswordHash() {
-        System.out.println(md5StringHasher);
-        String ab = md5StringHasher.getHash(this.password);
-        System.out.println(ab);
-        return ab;
+        return password;
     }
 
     @Override
@@ -55,7 +57,7 @@ public class LoginInstanceImpl implements LoginInstance {
 
     @Override
     public void setPassword(String password) {
-        this.password = password;
+        this.password = md5StringHasher.getHash(password);
     }
 
     @Override
