+++ src/test/java/MongoTests/LoginValidCheckerTests.java
@@ -1,29 +1,53 @@
-package LoginValidCheckerTests;
+package MongoTests;
+import com.mongodb.MongoClient;
+import com.mongodb.MongoClientURI;
+import com.mongodb.client.MongoDatabase;
 import org.junit.Before;
+import org.junit.BeforeClass;
+import org.junit.Rule;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.context.annotation.AnnotationConfigApplicationContext;
 import org.springframework.test.context.ContextConfiguration;
 import org.springframework.test.context.junit4.SpringRunner;
+import org.testcontainers.containers.GenericContainer;
+import org.testcontainers.containers.MongoDBContainer;
+import org.testcontainers.utility.DockerImageName;
 import ru.eatthefrog.hatterBot.LoginManager.LoginInstance;
+import ru.eatthefrog.hatterBot.LoginManager.LoginInstanceImpl;
 import ru.eatthefrog.hatterBot.LoginManager.LoginValidChecker;
 import ru.eatthefrog.hatterBot.MongoDBOperator.DataBaseLoginManager;
+import ru.eatthefrog.hatterBot.SpringConfiguration;
 
 import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.Matchers.equalTo;
 
 @RunWith(SpringRunner.class)
-@ContextConfiguration(classes = LoginValidCheckerTestsConfiguration.class)
+@ContextConfiguration(classes = SpringConfiguration.class)
 public class LoginValidCheckerTests {
+    @Rule
+    public GenericContainer mongoDBContainer = new GenericContainer(DockerImageName.parse("mongo:4.4.2"))
+            .withExposedPorts(27017);
+
+    @Autowired
+    MongoDatabase mongoDatabase;
+
     @Autowired
     LoginValidChecker loginValidChecker;
 
     @Autowired
     DataBaseLoginManager dataBaseLoginManager;
 
-    @Before
-    public void resetDatabase() {
-        dataBaseLoginManager.resetDatabase();
+    public static LoginInstance getLoginInstance(String login, String password) {
+        AnnotationConfigApplicationContext context = new AnnotationConfigApplicationContext(
+                SpringConfiguration.class
+        );
+        LoginInstance loginInstance = context.getBean("loginInstanceBean", LoginInstance.class);
+        context.close();
+        loginInstance.setLogin(login);
+        loginInstance.setPassword(password);
+        return loginInstance;
     }
 
     @Test
@@ -33,9 +57,9 @@ public class LoginValidCheckerTests {
         // Обращение к базе данных происходит лишь в двух случаях:
         // 1) В момент залогинивания или регистрации.
         // 2) В момент устаревания валидного loginInstance.
-        LoginInstance loginInstance1 = TestLoginInstanceFactory.getLoginInstance("abc","de");
+        LoginInstance loginInstance1 = getLoginInstance("Vasya","Hahaha");
         loginValidChecker.rememberLoginInDB(loginInstance1);
-        LoginInstance loginInstance2 = TestLoginInstanceFactory.getLoginInstance("abc","de");
+        LoginInstance loginInstance2 = getLoginInstance("Vasya","Hahaha");
         assertThat(loginValidChecker.checkValidLogin(loginInstance1), equalTo(true));
         assertThat(loginValidChecker.checkValidLogin(loginInstance2), equalTo(false));
         System.out.println(dataBaseLoginManager.getCount());
@@ -43,8 +67,8 @@ public class LoginValidCheckerTests {
     }
     @Test
     public void checkLoginInstanceInDB() {
-        LoginInstance loginInstance1 = TestLoginInstanceFactory.getLoginInstance("abc","de");
-        LoginInstance loginInstance2 = TestLoginInstanceFactory.getLoginInstance("abc","de");
+        LoginInstance loginInstance1 = getLoginInstance("abc","de");
+        LoginInstance loginInstance2 = getLoginInstance("abc","de");
         loginValidChecker.rememberLoginInDB(loginInstance1);
         assertThat(loginValidChecker.checkValidLogin(loginInstance1), equalTo(true));
         assertThat(loginValidChecker.checkValidLogin(loginInstance2), equalTo(false));
@@ -57,7 +81,7 @@ public class LoginValidCheckerTests {
 
     @Test
     public void notValidWithoutDB() {
-        LoginInstance loginInstance1 = TestLoginInstanceFactory.getLoginInstance("abc","de");
+        LoginInstance loginInstance1 = getLoginInstance("abc","de");
         loginValidChecker.rememberLoginInDB(loginInstance1);
         loginInstance1.setNotValid();
         assertThat(loginValidChecker.checkValidLogin(loginInstance1), equalTo(false));
@@ -69,7 +93,7 @@ public class LoginValidCheckerTests {
         // loginInstance, кладём его в БД. После чего от устаревает,
         // в то время база данных очистилась, мы верифицируем его.
         // Он оказывается неверным.
-        LoginInstance loginInstance1 = TestLoginInstanceFactory.getLoginInstance("abc","de");
+        LoginInstance loginInstance1 = getLoginInstance("abc","de");
         loginValidChecker.rememberLoginInDB(loginInstance1);
         assertThat(loginValidChecker.checkValidLogin(loginInstance1), equalTo(true));
         loginInstance1.makeUltraOld();
@@ -79,7 +103,7 @@ public class LoginValidCheckerTests {
 
     @Test
     public void oldLoginInstanceInDB() {
-        LoginInstance loginInstance1 = TestLoginInstanceFactory.getLoginInstance("abc","de");
+        LoginInstance loginInstance1 = getLoginInstance("abc","de");
         loginValidChecker.rememberLoginInDB(loginInstance1);
         assertThat(loginValidChecker.checkValidLogin(loginInstance1), equalTo(true));
         loginInstance1.makeUltraOld();
